+++ src/app/shared/components/basket/basket-cost-summary/basket-cost-summary.component.ts
@@ -1,6 +1,10 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
+import { AccountFacade } from 'ish-core/facades/account.facade';
 import { BasketTotal } from 'ish-core/models/basket-total/basket-total.model';
+import { PriceItemHelper } from 'ish-core/models/price-item/price-item.helper';
 import { PriceHelper } from 'ish-core/models/price/price.model';
 
 /**
@@ -16,8 +20,22 @@ import { PriceHelper } from 'ish-core/models/price/price.model';
   templateUrl: './basket-cost-summary.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class BasketCostSummaryComponent {
+export class BasketCostSummaryComponent implements OnInit {
   @Input() totals: BasketTotal;
 
+  taxTranslation$: Observable<string>;
   invert = PriceHelper.invert;
+
+  constructor(private accountFacade: AccountFacade) {}
+
+  ngOnInit() {
+    this.taxTranslation$ = this.accountFacade.userPriceDisplayType$.pipe(
+      map(type => (type === 'net' ? 'checkout.tax.text' : 'checkout.tax.TaxesLabel.TotalOrderVat'))
+    );
+  }
+
+  get hasPaymentCostsTotal(): boolean {
+    const paymentCosts = PriceItemHelper.selectType(this.totals && this.totals.paymentCostsTotal, 'gross');
+    return !!paymentCosts && !!paymentCosts.value;
+  }
 }